Change the product code to be a uint8_t consistently everwhere.
tic_settings_get_product was in a tagged release and it uses uint8_t, so we
should not change its signature and it's better to make everything else be
consistent with it.

tic_device_get_product has not been used in a tagged release yet, so it is OK to
change its signature.