nix/builder.sh: Remove 'lib' and 'include' directories.
This Nix recipe is for building the Tic software, not for building libpololu-tic
as a standalone thing.  If you want to link against libpololu-tic, you should
probably build and install the library with the native build tools of your
system, or use the tic recipe in nixcrpkgs (which does not exist yet).